Move bindings handling to EditWithGeneratedProps
