+++ packages/block-editor/src/hooks/use-bindings-attributes.js
@@ -1,18 +1,3 @@
-/**
- * WordPress dependencies
- */
-import { store as blocksStore } from '@wordpress/blocks';
-import { createHigherOrderComponent } from '@wordpress/compose';
-import { useRegistry, useSelect } from '@wordpress/data';
-import { useCallback, useMemo, useContext } from '@wordpress/element';
-
-/**
- * Internal dependencies
- */
-import isURLLike from '../components/link-control/is-url-like';
-import { unlock } from '../lock-unlock';
-import BlockContext from '../components/block-context';
-
 /** @typedef {import('@wordpress/compose').WPHigherOrderComponent} WPHigherOrderComponent */
 /** @typedef {import('@wordpress/blocks').WPBlockSettings} WPBlockSettings */
 
@@ -30,7 +15,7 @@ const BLOCK_BINDINGS_ALLOWED_BLOCKS = {
 	'core/button': [ 'url', 'text', 'linkTarget', 'rel' ],
 };
 
-const DEFAULT_ATTRIBUTE = '__default';
+export const DEFAULT_ATTRIBUTE = '__default';
 
 /**
  * Returns the bindings with the `__default` binding for pattern overrides
@@ -44,7 +29,7 @@ const DEFAULT_ATTRIBUTE = '__default';
  *
  * @return {Object} The bindings with default replaced for pattern overrides.
  */
-function replacePatternOverrideDefaultBindings( blockName, bindings ) {
+export function replacePatternOverrideDefaultBindings( blockName, bindings ) {
 	// The `__default` binding currently only works for pattern overrides.
 	if (
 		bindings?.[ DEFAULT_ATTRIBUTE ]?.source === 'core/pattern-overrides'
@@ -94,213 +79,3 @@ export function canBindAttribute( blockName, attributeName ) {
 export function getBindableAttributes( blockName ) {
 	return BLOCK_BINDINGS_ALLOWED_BLOCKS[ blockName ];
 }
-
-export const withBlockBindingSupport = createHigherOrderComponent(
-	( BlockEdit ) => ( props ) => {
-		const registry = useRegistry();
-		const blockContext = useContext( BlockContext );
-		const sources = useSelect( ( select ) =>
-			unlock( select( blocksStore ) ).getAllBlockBindingsSources()
-		);
-		const { name, clientId, context, attributes, setAttributes } = props;
-		const hasPatternOverridesDefaultBinding =
-			attributes?.metadata?.bindings?.[ DEFAULT_ATTRIBUTE ]?.source ===
-			'core/pattern-overrides';
-		const blockBindings = useMemo(
-			() =>
-				replacePatternOverrideDefaultBindings(
-					name,
-					attributes?.metadata?.bindings
-				),
-			[ attributes?.metadata?.bindings, name ]
-		);
-
-		const { computedAttributes, computedContext } = useSelect(
-			( select ) => {
-				if ( ! blockBindings ) {
-					return {
-						computedAttributes: attributes,
-						computedContext: context,
-					};
-				}
-
-				const attributesFromSources = {};
-				const contextFromSources = {};
-				const blockBindingsBySource = new Map();
-
-				for ( const [ attributeName, binding ] of Object.entries(
-					blockBindings
-				) ) {
-					const { source: sourceName, args: sourceArgs } = binding;
-					const source = sources[ sourceName ];
-					if (
-						! source ||
-						! canBindAttribute( name, attributeName )
-					) {
-						continue;
-					}
-
-					// Populate context.
-					for ( const key of source.usesContext || [] ) {
-						contextFromSources[ key ] = blockContext[ key ];
-					}
-
-					blockBindingsBySource.set( source, {
-						...blockBindingsBySource.get( source ),
-						[ attributeName ]: {
-							args: sourceArgs,
-						},
-					} );
-				}
-
-				if ( blockBindingsBySource.size ) {
-					for ( const [
-						source,
-						bindings,
-					] of blockBindingsBySource ) {
-						// Get values in batch if the source supports it.
-						let values = {};
-						if ( ! source.getValues ) {
-							Object.keys( bindings ).forEach( ( attr ) => {
-								// Default to the the source label when `getValues` doesn't exist.
-								values[ attr ] = source.label;
-							} );
-						} else {
-							values = source.getValues( {
-								select,
-								context: contextFromSources,
-								clientId,
-								bindings,
-							} );
-						}
-						for ( const [ attributeName, value ] of Object.entries(
-							values
-						) ) {
-							if (
-								attributeName === 'url' &&
-								( ! value || ! isURLLike( value ) )
-							) {
-								// Return null if value is not a valid URL.
-								attributesFromSources[ attributeName ] = null;
-							} else {
-								attributesFromSources[ attributeName ] = value;
-							}
-						}
-					}
-				}
-
-				return {
-					computedAttributes: {
-						...attributes,
-						...attributesFromSources,
-					},
-					computedContext: { ...context, ...contextFromSources },
-				};
-			},
-			[
-				attributes,
-				blockBindings,
-				blockContext,
-				clientId,
-				context,
-				name,
-				sources,
-			]
-		);
-
-		const setBoundAttributes = useCallback(
-			( nextAttributes ) => {
-				if ( ! blockBindings ) {
-					setAttributes( nextAttributes );
-					return;
-				}
-
-				registry.batch( () => {
-					const keptAttributes = { ...nextAttributes };
-					const blockBindingsBySource = new Map();
-
-					// Loop only over the updated attributes to avoid modifying the bound ones that haven't changed.
-					for ( const [ attributeName, newValue ] of Object.entries(
-						keptAttributes
-					) ) {
-						if (
-							! blockBindings[ attributeName ] ||
-							! canBindAttribute( name, attributeName )
-						) {
-							continue;
-						}
-
-						const binding = blockBindings[ attributeName ];
-						const source = sources[ binding?.source ];
-						if ( ! source?.setValues ) {
-							continue;
-						}
-						blockBindingsBySource.set( source, {
-							...blockBindingsBySource.get( source ),
-							[ attributeName ]: {
-								args: binding.args,
-								newValue,
-							},
-						} );
-						delete keptAttributes[ attributeName ];
-					}
-
-					if ( blockBindingsBySource.size ) {
-						for ( const [
-							source,
-							bindings,
-						] of blockBindingsBySource ) {
-							source.setValues( {
-								select: registry.select,
-								dispatch: registry.dispatch,
-								context: computedContext,
-								clientId,
-								bindings,
-							} );
-						}
-					}
-
-					const hasParentPattern =
-						!! computedContext[ 'pattern/overrides' ];
-
-					if (
-						// Don't update non-connected attributes if the block is using pattern overrides
-						// and the editing is happening while overriding the pattern (not editing the original).
-						! (
-							hasPatternOverridesDefaultBinding &&
-							hasParentPattern
-						) &&
-						Object.keys( keptAttributes ).length
-					) {
-						// Don't update caption and href until they are supported.
-						if ( hasPatternOverridesDefaultBinding ) {
-							delete keptAttributes?.caption;
-							delete keptAttributes?.href;
-						}
-						setAttributes( keptAttributes );
-					}
-				} );
-			},
-			[
-				blockBindings,
-				clientId,
-				computedContext,
-				hasPatternOverridesDefaultBinding,
-				setAttributes,
-				sources,
-				name,
-				registry,
-			]
-		);
-
-		return (
-			<BlockEdit
-				{ ...props }
-				context={ computedContext }
-				attributes={ computedAttributes }
-				setAttributes={ setBoundAttributes }
-			/>
-		);
-	},
-	'withBlockBindingSupport'
-);
